requirements.txt
Django>=4.2
djangorestframework>=3.15
django-filter>=23.2
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
